refactor outcome review detail router
