+++ src/app/router_registry.py
@@ -74,6 +74,9 @@ from app.routers.dpm_command_center_mandates import (
 from app.routers.dpm_command_center_monitoring import (
     router as dpm_command_center_monitoring_router,
 )
+from app.routers.dpm_command_center_outcome_review_detail import (
+    router as dpm_command_center_outcome_review_detail_router,
+)
 from app.routers.dpm_command_center_outcome_review_evidence import (
     router as dpm_command_center_outcome_review_evidence_router,
 )
@@ -373,6 +376,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_exceptions_router)
     app.include_router(dpm_command_center_exception_ai_router)
     app.include_router(dpm_command_center_outcome_reviews_router)
+    app.include_router(dpm_command_center_outcome_review_detail_router)
     app.include_router(dpm_command_center_outcome_review_lookup_router)
     app.include_router(dpm_command_center_outcome_review_evidence_router)
     app.include_router(dpm_command_center_outcome_review_handoff_router)
